Introduce QEMU_GA_ARGS env variable to replace BLACKLIST_RPC
We previously replaced use of --blacklist with -b since the former
was deleted in QEMU 9.1.

We want to avoid use of the 'black list' term downstream too though.
In future we're also expecting further command line args to be added
to the guest agent which are desirable to set.

Rather than hardcode the use of -b in the unit file, define a more
general QEMU_GA_ARGS variable that enables the sysconfig file to
set arbitrary extra command line arguments.

A pre-start script is used to rewrite existing sysconfig files that
users may have present to ensure upgradability.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>