spec: cherry pick openat2 qemu-user support from usptream 9651cea
This commit adds support for the openat2 syscall to qemu-user. It
is done via cherry picking upstream 9651cea and adding a extra
commit with a bunch of `#ifdef TARGET_NR_openat2` so that this
commit compiles on the `cris-linux-user` target which does not
have this syscall. Cris is removed in upstream qemu after v9.1.0
so the ifdefs were not needed there but are needed here until
cris is also removed from the RPM.