Fix riscv64 build
It currently fails with:

  RPM build errors:
  error: File not found: .../usr/lib/binfmt.d/qemu-riscv32-static.conf

This is caused by the qemu-binfmt-conf.sh script assuming that
all riscv64 CPUs can natively run riscv32 applications, which
is actually not the case in practice.

Backport a few patches from upstream. Once those have been
applied, the script can be convinced to behave correctly, i.e.
generate the file that until now was missing.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>