Drop have_edk2 global
edk2 binary packages are noarch so they can be installed on
mismatching architectures, e.g. edk2-aarch64 on x86_64.

So the existing conditionals are not only unnecessary, but
actively harmful, as they result in edk2-ovmf not being
dragged in automatically when qemu-system-x86-core is
installed on e.g. ppc64le.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>